添加 基础釉，色料 ， 客户 删除检测

+++ src/main/java/com/easygame/sdk/service/impl/backend/general/CompanyBizImpl.java
@@ -46,9 +46,11 @@ public class CompanyBizImpl extends BaseBizImpl implements ICompanyBiz {
 	public int deleteCompanyByPrimaryKey(int id) {
 		
 		if (companyMapper.checkCompanyDeleteSecurity(id) > 0) {
-			
 			return 0;
-			
+		}
+
+		if(companyMapper.checkCompanyUsedBySampleGlaze(id)> 0){
+			return 0;
 		}
 		
 		return companyMapper.deleteCompanyByPrimaryKey(id);
